Auto stash before merge of "develop-grind" and "origin/develop-grind"

+++ PlayMode.cpp
@@ -22,7 +22,7 @@ PlayMode::PlayMode() {
 	{ //walls
 		for (int i = 0; i < 30; i++)
 		{
-			float x = ((float)i * 17);
+			float x = (i * 17.f);
 			if (int(x) % 2 == 0) x = -x;
 			float y = (float)240 * (float)i / 30;
 			walls_at.push_back(glm::vec2(uint8_t(x) % 256, uint8_t(y) % 240));
